sizeWithFont is deprecated, using boundingRectWithSize instead
Getting rid of the sizeWithFont warning.

I did it in a way that I get the same behavior with long texts and that's why I used the whole frame size, to avoid people upgrading and getting different results with the badge. Feel free to add the max size on boundingRectWithSize:CGSize.